fuzzing/pyproject.toml
[tool.poetry]
name = "mutagen-fuzzing"
version = "0.1.0"
description = ""
authors = ["Christoph Reiter <reiter.christoph@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
cython = "^0.29.15"
python-afl = "^0.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
